pyproject.toml
[project]
name = "adapter_agent"
version = "0.1.0"
description = "OpenHands-quality coding agent using openai-agents-sdk"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "mcp>=1.25.0",
    "peft>=0.18.1",
    "polars>=1.37.1",
    "rich>=14.2.0",
    "transformers>=4.56.0",
    "trl>=0.26.2",
    "flash-attn",
    "vllm>=0.13.0",
    "litellm>=1.85.0.dev2",
    "openai>=2.14.0",
    "chz>=0.4.0",
    "tinker-cookbook",
    "oai-utils",
    "wandb>=0.24.1",
    "tinker>=0.9.0",
    "seaborn>=0.13.2",
    "weave>=0.52.28",
    "more-itertools>=10.8.0",
    "torch>=2.9.0",
    "coder-mcp",
    "ray[default]>=2.54.0",
    "debugpy>=1.8.20",
    "requests>=2.32.5",
    "urllib3>=2.6.3",
    "chardet~=5.0",
    "pyvis>=0.3.2",
    "networkx>=3.6.1",
    "streamlit>=1.55.0",
    "elasticsearch<9.0.0",
    "fastapi>=0.135.2",
    "uvicorn>=0.42.0",
    "aiosqlite>=0.22.1",
    "asyncpg>=0.29.0",
    "prisma>=0.15.0",
    "openai-agents>=0.14.1",
    "orjson>=3.11.8",
    "fastmcp>=2.2.0",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
no-binary-package = ["flash-attn"]
no-build-isolation-package = ["flash-attn"]
index-strategy = "unsafe-best-match"

[tool.uv.sources]
vllm = [
    { index = "vllm-cu130" }
]
tinker-cookbook = { git = "https://github.com/thinking-machines-lab/tinker-cookbook" }
oai-utils = { path = "../oai_utils", editable = true }
coder-mcp = { path = "../coder-mcp", editable = true }
[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]



[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[[tool.uv.index]]
name = "vllm-cu130"
url = "https://wheels.vllm.ai/0.13.0/cu130"
explicit = true
